# openapi.yaml
openapi: "3.1.0"
info:
  title: Get-a-J*b API
  description: |
    Job application tracker API. Manage jobs, resumes, and users
    with JWT-based authentication and refresh token rotation.
  version: "1.0.0"
  contact:
    name: ShoAnn
servers:
  - url: http://localhost:8080
    description: Local development server
paths:
...
  /api/jobs:
    post:
      operationId: createJob
      summary: Create a new job entry
      tags: [Jobs]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateJobRequest"
      responses:
        "201":
          description: Job created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
    get:
      operationId: getAllJobs
      summary: List all jobs for the authenticated user
      tags: [Jobs]
      security:
        - BearerAuth: []
      responses:
        "200":
          description: List of jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Job"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
  /api/jobs/{id}:
    get:
      operationId: getJobById
      summary: Get a job by ID
      tags: [Jobs]
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Job details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "400":
          description: Invalid job ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                message: Invalid input
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                message: Forbidden
        "404":
          description: Job not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                message: Job not found
    put:
      operationId: updateJob
      summary: Update a job by ID
      tags: [Jobs]
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateJobRequest"
      responses:
        "200":
          description: Job updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          description: Job not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                message: Job not found
    delete:
      operationId: deleteJob
      summary: Delete a job by ID
      tags: [Jobs]
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Job deleted successfully
        "400":
          description: Invalid job ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                message: Invalid input
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          description: Job not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                message: Job not found
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token (JWT, 24-hour expiry)
  schemas:
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          description: Error message
      required: [message]
    ValidationError:
      type: object
      properties:
        errors:
          type: object
          additionalProperties:
            type: string
          description: Field-level validation error messages
          example:
            email: must be a valid email address
            username: this field is required
      required: [errors]
...
    Job:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: integer
          description: Owner user ID
        title:
          type: string
        company:
          type: string
        location:
          type: string
        salary:
          type: integer
        description:
          type: string
        requirements:
          type: string
        current_status:
          type: string
          enum: [draft, submitted, under_review, interview_scheduled, offer_extended, accepted, rejected, withdrawn, archived]
        status_changed_at:
          type: string
          format: date-time
        notes:
          type: string
        source_url:
          type: string
          format: uri
        job_portal:
          type: string
        created_at:
          type: string
          format: date-time
      required: [id, user_id, title, company, location, salary, description, requirements, current_status, status_changed_at, notes, source_url, job_portal, created_at]
    CreateJobRequest:
      type: object
      required: [title, company, location, salary, requirements]
      properties:
        title:
          type: string
        company:
          type: string
        location:
          type: string
        salary:
          type: integer
        description:
          type: string
        requirements:
          type: string
        current_status:
          type: string
          enum: [draft, submitted, under_review, interview_scheduled, offer_extended, accepted, rejected, withdrawn, archived]
          default: draft
        notes:
          type: string
        source_url:
          type: string
          format: uri
        job_portal:
          type: string
    UpdateJobRequest:
      type: object
      properties:
        title:
          type: string
        company:
          type: string
        location:
          type: string
        salary:
          type: integer
        description:
          type: string
        requirements:
          type: string
        source_url:
          type: string
          format: uri
        current_status:
          type: string
          enum: [draft, submitted, under_review, interview_scheduled, offer_extended, accepted, rejected, withdrawn, archived]
        notes:
          type: string
        job_portal:
          type: string
...
  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationError"
    UnauthorizedError:
      description: Missing or invalid JWT token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            message: Unauthorized
...